refactor(web): move annual page to (home) group
