+++ apps/web/src/app/(dashboard)/(home)/annual/page.tsx
@@ -1,4 +1,4 @@
-import { AnnualRegistrationsChart } from "@web/app/(dashboard)/annual/_components/annual-registrations-chart";
+import { AnnualRegistrationsChart } from "@web/app/(dashboard)/(home)/annual/_components/annual-registrations-chart";
 import { StructuredData } from "@web/components/structured-data";
 import Typography from "@web/components/typography";
 import { SITE_TITLE, SITE_URL } from "@web/config";
@@ -45,7 +45,7 @@ const AnnualPage = async () => {
     <>
       <StructuredData data={structuredData} />
       <section className="flex flex-col gap-8">
-        <div className="flex flex-col gap-2">
+        <div className="flex flex-col">
           <Typography.H1>Annual Registrations</Typography.H1>
           <Typography.TextLg className="text-default-500">
             Total new car registrations in Singapore by year
